Claritfy nature of unmarked status
Why:
Up to now there were 2 ways to have an unmarked status. Either by the
field not being at all specified, or by the empty string. This was
confusing and was necessitating needlessly more logic.

What:
Switch to always requiring presence, leaving the empty string for the
unmarked status. Update tests to reflect that