Pad accounts and amounts properly
Why:
We want our journal files to be somewhat pleasing to the eye, at least
in the vertical direction, in order to allow easier inspection

What:
Solve the proper alignment TODO we had regarding posting account and
amounts by padding the postings with the shorter string lengths with
enough spaces. Pad the amounts as well to make sure numbers are
vertically aligned, regardless of sign. We now no longer use only 2
spaces to separate amount from account, but rather 3 in the positive
case and 2 in the negative sign case

Update tests as well to reflect all that