Add MyExpenses JSON importer
Why:
That's the entire point of the project. To consume MyExpenses exports
and import them. The most easy one to work with, as in the least
ambiguous is the JSON one

What:
Add in a few functions to read the JSON export and transform it to what
hledger awaits

Import data.json as it is needed for this to work