Improve formatting
Why:
There were a number of shortcomings with formatting transaction. Namely:

Balancing transaction would would not be properly aligned on the last
digit of the integer part, which we want cause it makes easier to
inspect the ledger

Separate transactions would have different alignments for their
postings. This also made it a bit more difficult to visually inspect the
ledger

What:
* Move integer part splitting into dedicated, tested function
* Split out the max length calculation part into dedicated, tested
  function
* Make the format-transaction function double-arity to support the case
  of caring as well as the case for not caring to properly align a
  transaction as the rest of them. Simplify the function in the process
  using the above ones
* Rename bindings to make their intent clearer

For now, disable the tests regarding costs, it's pointless to have it
around until we do support it