Log when a transaction fails to conform to spec
Why:
Having an audit log of what is going on and why some transactions fail
to be import (e.g. their status is void in MyExpenses) is pretty useful

What:
Move the conform check a bit higher and then log an explain-data stanza.
For now, just log the first problem