Explicitly output transaction strings
Why:
The previous version was a bit implicit, redirecting *out* to a string
and then using print commands. This is a bit more difficult to grok.
Furthermore print form injects spaces between arguments, leading to
unnecessary whitespace

What:
Switch to format instead of print and just join strings on "\n"