Implementing processing a directory
Why:
We want to also to be able to traverse a directory of JSON files as it
is possible to choose this mode of export in MyExpenses, by either
selectively export accounts one by one, or exporting a summary (e.g. EUR
one) account as individual accounts.

What:
Traverse files in a directory. The burden of not messing up that
directory and only have JSON files from MyExpenses in there is left up
to the user, but will be documented