Only store dedup info if transaction conforms to spec
Why:
If a transaction does not conform to spec, we aren't going to be seeing
it anyway in the output. If there is a 2nd MyExpenses transaction with
the same UUID, which does conform to the spec, we 'll end up having the
plain text accounting transaction we want anyway. Effectively, we are
increasing our chances of properly converting the transaction

What:
Move the storing of the "code" part of the transaction in the else
clause of the if stanza that checks for conformity of transaction